+++ src/components/Delivery/UserDetails/index.js
@@ -4,7 +4,7 @@ import Skeleton from 'react-loading-skeleton'
 import { useTheme } from 'styled-components'
 import { Dropdown, DropdownButton } from 'react-bootstrap'
 import { useLanguage, UserDetails as UserDetailsController } from 'ordering-components-admin'
-import { OriginalOrdersManager as OrdersManager } from '../../Orders/OrdersManager/Layouts/OriginalOrdersManager'
+import { OrdersManager } from '../../Orders/OrdersManager'
 import { AddressList } from '../AddressList'
 import { Personalization, Modal, Schedule } from '../../Shared'
 import { UserDetailsMenu } from '../UserDetailsMenu'
