+++ template/theme.json
@@ -58,12 +58,5 @@
   },
   "defaultLanguages": {
     "LOADING_DELICIOUS_FOOD": "Loading Ordering Dashboard..."
-  },
-  "orders_manager_view": {
-    "components": {
-      "layout": {
-        "type": "original"
-      }
-    }
   }
 }
